requirements.txt
streamlit==1.49.1
numpy==2.3.2
openai==1.12.0
requests==2.32.5
pydrive==1.3.1
python-dotenv==1.0.1
oauth2client==4.1.3
sqlalchemy==2.0.27
pdfminer.six==20231228
streamlit_lottie==0.0.5
psycopg2-binary==2.9.9
google-cloud-aiplatform==1.43.0
